[BOOK-281] feat: card에서 필요한 VO 파일 확장
